// frontend/new-web/src/components/Crud/types.ts
export type IObject = Record<string, any>;

/**
 * 组件类型定义
 */
type DateComponent = "date-picker" | "time-picker" | "time-select" | "custom-tag" | "input-tag";
type InputComponent = "input" | "select" | "input-number" | "cascader" | "tree-select";
type OtherComponent =
  | "text"
  | "radio"
  | "checkbox"
  | "switch"
  | "rate"
  | "slider"
  | "icon-select"
  | "custom";
export type ISearchComponent =
  | DateComponent
  | InputComponent
  | "radio"
  | "checkbox"
  | "switch"
  | "rate"
  | "slider"
  /** 在 customComponents 中注册的自定义搜索控件类型 */
  | "user-table-select";
export type IComponentType = DateComponent | InputComponent | OtherComponent;
// ...
export type IFormItems<T = IComponentType> = Array<{
  /** 组件类型(如input,select,radio,custom等) */
  type: T;
  /** 标签提示 */
  tips?: string | IObject;
  /** 标签 */
  label: string;
  /** 属性名 */
  prop: string;
  /** 属性 */
  attrs?: IObject;
  /** 选项 */
  options?: Array<{ label: string; value: any; [key: string]: any }> | Ref<any[]>;
  /** 规则 */
  rules?: FormItemRule[];
  /** 初始值 */
  initialValue?: any;
  /** 插槽名称 */
  slotName?: string;
  /** 是否隐藏 */
  hidden?: boolean;
  /** 列属性 */
  col?: Partial<ColProps>;
  /** 事件 */
  events?: Record<string, (...args: any) => void>;
  /** 初始化函数 */
  initFn?: (item: IObject) => void;
}>;